parted: don't reload partition table on every command
gpt was using a static local variable to suppress repeatedly reporting
an error if you chose to ignore it.  This is incorrect as the variable is
global to all disks, and ignoring the error on one should not suppress its
reporting on another.  Moving the flag to the PedDisk object made it
effectively useless because parted was destroying the PedDisk and reloading
the partition table on every command.

Parted has been reworked to cache the PedDisk once loaded, and only discard
it when changing disks, or creating a new disklabel.