libparted: mac: support sector size > 512 correctly
This patch fixes several problems found when trying to do Apple
Partition Map on a 4K-sector iPod. The problems happened because of
unneeded conversion between physical sectors and 512-byte units.
For devices with sector 512 the conversion did no harm, for others it did.

List of problems (all are gone with the patch):

* Partition in OS X, connect to Linux, run "parted <dev> print".
  "Error: Can't have a partition outside the disk!" is shown.
  Every partition is displayed 8 times (4k/512) bigger its real size.

* Exec "mktable mac", connect to OS X.
  The drive's partitioning scheme is not recognized.

* mkpart creates partitions 8 times (4k/512) smaller than requested.

Signed-off-by: Sergei Antonov <saproj@gmail.com>