libparted: Fix check for backup header location
Add a couple of helper functions for calculating the partition table
entry size (in sectors) and for guessing the end of the disk based on
the LastUsableLBA and the Partition Table Entry size.

The backup header should be either at the end of the disk, or at what
the primary header thinks is the end of the disk. Prompt to fix the
backup header if it is located any other place.

* libparted/labels/gpt.c (_ptes_sectors): New function
  (_hdr_disk_end): New function
  (gpt_read): Use new function to test for pri's idea of end of disk