tests: fix t1700
bb181a7e: "tests: Use force for xfs in t1700 and a larger file" caused
the previous filesystem signatures to be left in the image file
causing mkfs.nilfs2 to complain that there is already an xfs filesystem
there, and hang the test suite waiting for an answer to proceed or not.
Remove the file between filesystems so it is recreated cleanly again.