libparted: don't trash filesystem when writing loop label
If you deleted the fake partition on a loop label, loop_write() would write
the loop signature to the device, zeroing out all other bytes in the first
sector.  When the disk contained an ext[234] filesystem and was using 2k
sectors, this would trash the super block residing in the 1-2kb part of the
sector causing the disk to become unrecognized.  Instead, read the existing
sector and only modify the first few bytes that contain the loop label.